refactor: 💡 modified sanitize-analytics helper function
This commit will modify sanitize-analytics helper function

BREAKING CHANGE: '
